requirements.txt
attrs==21.4.0
black==21.12b0
click==8.0.3
flake8==4.0.1
Flask==2.0.2
-e git+https://github.com/will-afs/CooldownManager.git@7a4ecf8f420fdad04af0afbdc6c370e7ebf9c166#egg=flaskapp
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
pytest==6.2.5
python-dotenv==0.19.2
singleton-decorator==1.0.0
toml==0.10.2
tomli==1.2.3
typing-extensions==4.0.1
Werkzeug==2.0.2
